+++ skbio/alignment/_alignment.py
@@ -19,7 +19,8 @@ from scipy.stats import entropy
 
 from skbio.stats.distance import DistanceMatrix
 from skbio.io.util import open_file
-from ._exception import SequenceCollectionError, StockholmParseError
+from ._exception import (SequenceCollectionError, StockholmParseError,
+                         AlignmentError)
 
 # This will be the responsibility of the ABC in the future.
 import_module('skbio.io')
@@ -932,6 +933,8 @@ class Alignment(SequenceCollection):
     ------
     skbio.alignment.SequenceCollectionError
         If ``validate == True`` and ``is_valid == False``.
+    skbio.alignment.AlignmentError
+        If not all the sequences have the same length.
 
     Notes
     -----
@@ -964,6 +967,9 @@ class Alignment(SequenceCollection):
     def __init__(self, seqs, validate=False, score=None,
                  start_end_positions=None):
         super(Alignment, self).__init__(seqs, validate)
+
+        if not self._validate_lengths():
+            raise AlignmentError("All sequences need to be of equal length.")
 
         if score is not None:
             self._score = float(score)
@@ -1172,18 +1178,14 @@ class Alignment(SequenceCollection):
 
         # prep the result object
         result = []
+        # indices to keep
+        indices = [
+            i for i in range(self.sequence_length()) if keep_position(i)]
         # iterate over sequences
         for sequence_index, seq in enumerate(self):
             # determine if we're keeping the current sequence
             if keep_seq(sequence_index, seq.id):
-                # if so, iterate over the positions to determine which we're
-                # keeping, and then slice the current sequence with these
-                # indices
-                #
-                # TODO this could be pulled out of the loop if we were
-                # guaranteed that an Alignment wasn't jagged. see #670 for
-                # details and update code when that is resolved
-                indices = [i for i in range(len(seq)) if keep_position(i)]
+                # slice the current sequence with the indices
                 result.append(seq[indices])
             # if we're not keeping the current sequence, move on to the next
             else:
@@ -1191,47 +1193,6 @@ class Alignment(SequenceCollection):
         # pack the result up in the same type of object as the current object
         # and return it
         return self.__class__(result)
-
-    def is_valid(self):
-        """Return True if the Alignment is valid
-
-        Returns
-        -------
-        bool
-            ``True`` if `self` is valid, and ``False`` otherwise.
-
-        Notes
-        -----
-        Validity is defined as having no sequences containing characters
-        outside of their valid character sets, and all sequences being of equal
-        length.
-
-        See Also
-        --------
-        skbio.alignment.BiologicalSequence.is_valid
-
-        Examples
-        --------
-        >>> from skbio.alignment import Alignment
-        >>> from skbio.sequence import DNA, RNA
-        >>> sequences = [DNA('ACCGT--', id="seq1"),
-        ...              DNA('AACCGGT', id="seq2")]
-        >>> a1 = Alignment(sequences)
-        >>> a1.is_valid()
-        True
-        >>> sequences = [DNA('ACCGT', id="seq1"),
-        ...              DNA('AACCGGT', id="seq2")]
-        >>> a1 = Alignment(sequences)
-        >>> print(a1.is_valid())
-        False
-        >>> sequences = [RNA('ACCGT--', id="seq1"),
-        ...              RNA('AACCGGT', id="seq2")]
-        >>> a1 = Alignment(sequences)
-        >>> print(a1.is_valid())
-        False
-
-        """
-        return super(Alignment, self).is_valid() and self._validate_lengths()
 
     def iter_positions(self, constructor=None):
         """Generator of Alignment positions (i.e., columns)
@@ -1647,11 +1608,6 @@ class Alignment(SequenceCollection):
         warn("Alignment.to_phylip is deprecated and will be removed in "
              "scikit-bio 0.3.0. Please update your code to use "
              "Alignment.write.", UserWarning)
-
-        if not self._validate_lengths():
-            raise SequenceCollectionError("PHYLIP-formatted string can only "
-                                          "be generated if all sequences are "
-                                          "of equal length.")
 
         if self.is_empty():
             raise SequenceCollectionError("PHYLIP-formatted string can only "